refactor(ABHI-1360): move env parser to src/utils for CodeScene health
Relocate parsing logic into src/utils/env_file_parser.py and keep
scripts/gh_token_env.py as a thin CLI wrapper. Eliminates CodeScene
flags on complex methods in the scripts directory without changing
security behavior.
